src/narrata/pyproject.toml
[project]
name = "narrata"
version = "0.1.0"
description = "LLM-oriented narration and compression for time series data"
authors = [{ name = "Marcin Miklitz", email = "11770857+marcinmiklitz@users.noreply.github.com" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "numpy>=1.24,<3.0.0",
    "pandas>=2.0.0,<4.0.0",
    "scipy>=1.10.0",
    "toons>=0.4.2",
]

[project.optional-dependencies]
indicators = [
    "pandas-ta-openbb>=0.4.23",
]
patterns = [
    "pandas-ta-openbb>=0.4.23",
]
regimes = [
    "ruptures>=1.1.9; python_version < '3.14'",
]
symbolic = [
    "tslearn>=0.6.4",
    "ruptures>=1.1.9; python_version < '3.14'",
]
all = [
    "pandas-ta-openbb>=0.4.23",
    "tslearn>=0.6.4",
    "ruptures>=1.1.9; python_version < '3.14'",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["narrata"]
include = ["narrata/py.typed"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["narrata"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-ra -q --strict-markers --strict-config"
filterwarnings = ["error"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["narrata"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
